centralização do controle da autenticação
