+++ src/components/layout.tsx
@@ -1,15 +1,18 @@
 import React, { useEffect, useCallback } from 'react';
-import { useDispatch } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 
 import { createMuiTheme, ThemeProvider } from '@material-ui/core/styles';
 import { red, blue } from '@material-ui/core/colors';
 import Container from '@material-ui/core/Container';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import useMediaQuery from '@material-ui/core/useMediaQuery';
+import Box from '@material-ui/core/Box';
 
 import Header from './header';
 import SEO from './seo';
+import useAuthCheck from '../hooks/useAuthCheck';
 
+import { AuthState } from '../models/AuthState';
 import { initProperties } from '../store/actions';
 
 const theme = createMuiTheme({
@@ -31,6 +34,7 @@ const Layout: React.FC<Props> = ({ title, children }) => {
     const dispatch = useDispatch();
     const isMobile = useMediaQuery(theme.breakpoints.down('sm'));
 
+    const authLoading = useSelector((state: AuthState) => state.auth.loading);
     const onInitProperties = useCallback(
         (isMobile: boolean) => dispatch(initProperties(isMobile)),
         [],
@@ -40,14 +44,22 @@ const Layout: React.FC<Props> = ({ title, children }) => {
         onInitProperties(isMobile);
     }, [onInitProperties, isMobile]);
 
-    return (
-        <ThemeProvider theme={theme}>
-            <CssBaseline />
-            <Header />
-            <SEO title={title} />
-            <Container maxWidth="md">{children}</Container>
-        </ThemeProvider>
-    );
+    useAuthCheck(location);
+
+    let conteudo = <div />;
+
+    if (!authLoading) {
+        conteudo = (
+            <Box>
+                <CssBaseline />
+                <Header />
+                <SEO title={title} />
+                <Container maxWidth="md">{children}</Container>
+            </Box>
+        );
+    }
+
+    return <ThemeProvider theme={theme}>{conteudo}</ThemeProvider>;
 };
 
 export default Layout;
